add aarch64-conda-linux-gnu support, simplify logic in build_env.sh, add ltrace and strace package output

+++ recipe/build_scripts/ltrace_target.sh
@@ -4,6 +4,11 @@ set -e
 
 . ${RECIPE_DIR}/build_scripts/build_env.sh
 
+# under construction ... until done disabled for arm
+if [ "${CFG_ARCH}" = "arm" ]; then
+  exit 0
+fi
+
 rm -rf "${WDIR}/build/ltrace-target"
 mkdir -p "${WDIR}/build/ltrace-target"
 pushd "${WDIR}/build/ltrace-target"
